+++ src/test_env.rs
@@ -173,9 +173,9 @@ impl StatsTrackerStorage for MockedStatsTrackerStorage {
         &self,
         _key_meta: &[(String, Value)],
         _update: &stats_tracker_storage::Aggregate,
-        _value: f64,
-    ) -> anyhow::Result<f64> {
-        Ok(0.0)
+        _value: i64,
+    ) -> anyhow::Result<i64> {
+        Ok(0)
     }
 }
