+++ src/tracker/mod.rs
@@ -23,8 +23,8 @@ impl StatsTracker {
         &self,
         key_meta: impl IntoIterator<Item = (String, Value)> + Send,
         aggregate: &stats_tracker_storage::Aggregate,
-        value: f64,
-    ) -> Result<f64> {
+        value: i64,
+    ) -> Result<i64> {
         let key_meta = key_meta.into_iter().collect::<Vec<_>>();
         self.store.update_aggr(&key_meta, aggregate, value).await
     }
